add Fenwick tree (binary indexed tree), sqrt-decomposition moved in separate file
